pyproject.toml
[project]
name = "dailyword-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot==13.11",
    "redis",
    "oauth2client",
    "gspread",
    "pandas",
    "pydrive",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-flake8",
    "pytest-mock",
    "flake8==4.0.1",
    "fakeredis",
    "autopep8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers"
norecursedirs = [
    "terraform",
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    "build",
    "dist",
]

